// server/interfaces/product/category.ts
// Generated by https://quicktype.io

export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>

export interface ListCategoryReturns {
  brand: Partial<Brand>
  tag_list: Partial<TagList[]>
  brand_list: Partial<BrandList[]>
  category_ad: Partial<CategoryAd>
  product_list: Partial<ProductList[]>
}

export interface Brand {
  name: string
  positioning: string
}

export interface TagList {
  tag_name: string
  tag_id: number
}

// Generated by https://quicktype.io

export interface BrandsProductsReturn {
  brand_list: BrandList[]
  category_ad: CategoryAd
  product_list: ProductList[]
}

export interface BrandList {
  brand_id: number
  name: string
  default_logo: string
  select_logo: string
}

export interface CategoryAd {
  image: string
  title: string
  type: string
  product_id: string
  tag_id: string
  skip_url: string
}

export interface ProductList {
  product_id: number
  tag: string
  image: string
  brand_id: number
  name: string
  sub_title: string
  property_tags: string[]
  price: number
  price_unit: string
  base_fee: string
  compare_buy: string
}